// src/types/department.ts
export type Department = {
    id: number;
    department_code: string;
    sale_office_id: number;
    description: string;
    group_code: string;
    ship_id: number;
    is_default: boolean;
    name_th: string;
    name_en: string;
    status: boolean;
    create_at?: string;
    update_at?: string;
};